+++ libs/extensions/angular/image-banner/src/image-banner.component.scss
@@ -6,24 +6,32 @@ $max-content-width: 324px;
 $main-image-height-small: 132px;
 $main-image-height-large: 164px;
 
-:host(.none) {
-  .blur-image {
-    display: none;
+// The breakpoint where container should switch from "narrow" to "wide"
+$container-query-breakpoint: 600px;
+
+@mixin container-less-than-width() {
+  @container banner (width < #{$container-query-breakpoint}) {
+    @content;
   }
+}
 
-  .main-content-body-action-link {
-    color: var(--kirby-semi-dark);
+@mixin container-more-than-width() {
+  @container banner (width >= #{$container-query-breakpoint}) {
+    @content;
   }
 }
 
-:host(.dark) .blur-image {
-  filter: blur($blur-radius) brightness(0.8);
+// Elements with container-type: inline-size will collapse on Safari when navigating which prevents Angular from setting the correct scroll position on return.
+// Until the issue is fixed on Safari we cannot use container queries. It's currently unknown whether this issue is being worked on or not.
+:host(:not(.container-more-than-width, .container-less-than-width)) {
+  display: block;
+  container-name: banner;
+  container-type: inline-size;
 }
 
-:host(.light) .blur-image {
-  filter: blur($blur-radius) brightness(1.3);
-}
-
+// Temporary styling to support desired behaviour without container queries.
+// See ImageBannerResizeDirective directive.
+// START
 :host(.container-less-than-width) {
   .dismiss {
     // on mobile we always want the button to be white as it sits on top of the image
@@ -79,6 +87,38 @@ $main-image-height-large: 164px;
   }
 }
 
+// Temporary styling to support desired behaviour without container queries.
+// See ImageBannerResizeDirective directive.
+// END
+
+:host(.none) {
+  .blur-image {
+    display: none;
+  }
+
+  .dismiss {
+    // on mobile we always want the button to be white as it sits on top of the image
+    @include container-less-than-width {
+      --kirby-inputs-background-color: var(--kirby-white);
+      --kirby-inputs-background-color-hover: var(--kirby-dark-overlay-10);
+      --kirby-inputs-background-color-active: var(--kirby-dark-overlay-20);
+      --kirby-inputs-color: var(--kirby-black);
+    }
+  }
+
+  .main-content-body-action-link {
+    color: var(--kirby-semi-dark);
+  }
+}
+
+:host(.dark) .blur-image {
+  filter: blur($blur-radius) brightness(0.8);
+}
+
+:host(.light) .blur-image {
+  filter: blur($blur-radius) brightness(1.3);
+}
+
 .blur-image-wrapper {
   // We scale the image here with inset to avoid halo-bleed when using blur.
   position: absolute;
@@ -94,22 +134,27 @@ $main-image-height-large: 164px;
   object-position: center;
 }
 
-kirby-card {
-  height: 100%;
-}
-
 .main-content-wrapper {
   width: 100%;
   padding: utils.size('xxs');
   box-sizing: border-box;
   display: flex;
   flex-direction: column;
+
+  @include container-more-than-width {
+    gap: utils.size('s');
+    flex-direction: initial;
+  }
 }
 
 .main-content-image-wrapper {
   display: flex;
   overflow: hidden;
   border-radius: utils.border-radius('s');
+
+  @include container-more-than-width {
+    flex: 1;
+  }
 }
 
 .main-content {
@@ -126,6 +171,10 @@ kirby-card {
 
   .main-content-header {
     padding-inline-end: utils.size('xxs');
+
+    @include container-more-than-width {
+      padding-inline-end: utils.size('xl');
+    }
   }
 
   &:has(.main-content-body-action-link) {
@@ -140,6 +189,12 @@ kirby-card {
 
   &:has(.body-text:empty):has(.header-text:empty) {
     padding-block-end: 0;
+  }
+
+  @include container-more-than-width {
+    flex: 1;
+    gap: utils.size('xs');
+    padding: utils.size('xxs') utils.size('xxs') utils.size('xxs') 0;
   }
 }
 
@@ -162,6 +217,10 @@ kirby-card {
   height: $main-image-height-small;
   object-fit: cover;
   object-position: center;
+
+  @include container-more-than-width {
+    height: $main-image-height-large;
+  }
 }
 
 .main-content-body {
@@ -169,6 +228,17 @@ kirby-card {
   justify-content: space-between;
   gap: utils.size('s');
   height: 100%;
+
+  @include container-more-than-width {
+    flex-direction: column;
+    max-width: $max-content-width;
+  }
+
+  .main-content-body-action-link {
+    @include container-more-than-width {
+      display: none;
+    }
+  }
 }
 
 .main-content-body-text {
@@ -176,10 +246,21 @@ kirby-card {
   overflow: hidden;
   max-height: utils.size('xl');
   padding-inline-end: utils.size('xxs');
+
+  @include container-more-than-width {
+    padding-inline-end: utils.size('xxl');
+    max-height: utils.size('xxxxl');
+  }
 }
 
 .main-content-body-action-text {
   display: none;
+
+  @include container-more-than-width {
+    align-self: start;
+    display: inline-flex;
+    margin: 0; // buttons have xxxs margin so reset that
+  }
 }
 
 .dismiss {
